Finalização do projeto de POO.

+++ src/entities/Avaliavel.java
@@ -2,9 +2,9 @@ package entities;
 
 // INTERFACE: Garante a obrigatoriedade das avaliações no sistema
 public interface Avaliavel {
-    void evaluar(int nota, String comentario);
+    void evaluar(double nota, String comentario);
 
-    void avaliar(int nota, String comentario);
-    int getNota();
+    void avaliar(double nota, String comentario);
+    double getNota();
     String getComentario();
 }
