+++ src/entities/JogoDigital.java
@@ -24,6 +24,6 @@ public class JogoDigital extends Jogo {
     }
 
     @Override
-    public void evaluar(int nota, String comentario) {
+    public void evaluar(double nota, String comentario) {
     }
 }
