+++ src/program/Programa.java
@@ -89,7 +89,7 @@ public class Programa {
                         System.out.print("\nDeseja dar uma nota e comentário agora? (S/N): ");
                         if (scanner.nextLine().equalsIgnoreCase("S")) {
                             System.out.print("Digite a nota de avaliação (0 a 5): ");
-                            int novaNota = Integer.parseInt(scanner.nextLine());
+                            double novaNota = Double.parseDouble(scanner.nextLine());
                             System.out.print("Escreva seu comentário/review: ");
                             String novoComent = scanner.nextLine();
                             jAlterar.avaliar(novaNota, novoComent);
